better jest structure between node and browser environments
