+++ packages/common/tests/setup/connection.ts
@@ -1,0 +1,2 @@
+const Connection = require('../../src/services/Connection')
+jest.mock('../../src/services/Connection')
